Rename headers to use .hpp, extension
