+++ rviz_plugin_tutorials/src/drive_widget.hpp
@@ -26,6 +26,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
+
 #ifndef DRIVE_WIDGET_H
 #define DRIVE_WIDGET_H
