+++ rviz_plugin_tutorials/src/teleop_panel.hpp
@@ -26,6 +26,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
+
 #ifndef TELEOP_PANEL_H
 #define TELEOP_PANEL_H
